Update Example
Included both map and list usage.
