+++ examples/HTTPS_StreamParser_Map/HTTPS_StreamParser_Map.ino
@@ -3,6 +3,8 @@
  * and takes the data directly from the WiFi receive buffers and begins parsing.
  *
  * The results we care about are then put into a stl::map for lookup based on key.
+ * All results are placed into a forward_list for printing as well.
+ *
  * Not totally memory efficient, but not the worst either.
  *
  * The use of SSL is by far the biggest consumer of memory, with over 20kB of SRAM consumed by this alone.
@@ -10,18 +12,20 @@
 #include <ESP8266WiFi.h>
 #include <WiFiClientSecureBearSSL.h>  
 #include <ESP8266HTTPClient.h>  
-#include "ExampleHandler.h"                 // Custom JSON document handler
+#include "WeatherForecastHandler.h"                 // Custom JSON document handler
 #include <ArduinoStreamParser.h>            // <==== THE JSON Streaming Parser - Arduino STREAM WRAPPER
 #include <map>
+#include <forward_list>
 #include <string>
+#include <TimeLib.h>  						// Requires TimeLib: https://github.com/PaulStoffregen/Time
 
 HTTPClient http; // Used for both SSl and non-SSL connections
 
 void printHeapFreeToSerial()
 {
-   Serial.print("Heap free:");
+   Serial.print("Heap free: ");
    Serial.println(String(ESP.getFreeHeap(), DEC));
-   Serial.print("Stack free:");
+   Serial.print("Stack free: ");
    Serial.println(String(ESP.getFreeContStack(), DEC));   
 }
 
@@ -38,7 +42,7 @@ void setup()
     
     Serial.println("Starting........."); printHeapFreeToSerial();
     
-    WiFi.begin("XXXXX", "XXXXX");
+    WiFi.begin("<AP SSID>", "<PASSWORD>");
     
     Serial.print("Connecting");
     while (WiFi.status() != WL_CONNECTED)
@@ -64,7 +68,7 @@ void setup()
 
     // Open Weather Map JSON Streaming Parser
     ArudinoStreamParser parser;
-    ExampleHandler custom_handler;
+    WeatherForecastHandler custom_handler;
     
     parser.setHandler(&custom_handler); // Link to customer listener (parser to be honest)
     
@@ -111,15 +115,44 @@ void setup()
     Serial.println("Done.");    
     printHeapFreeToSerial();   
     
-    Serial.println("----------------- PRINTING VALUES -----------------");
+    Serial.println("----------------- PRINTING MAP VALUES -----------------");
     for (auto& x: mymap) {
             Serial.print(x.first.c_str()); Serial.print (" = " ); Serial.println(x.second);
     }  
- 
+	
+   
 }
 
 
 void loop() {
   // put your main code here, to run repeatedly:
+
+    Serial.println("----------------- PRINTING FORECAST LIST -----------------");
+    for (auto& x: myForecasts) {
+
+      // Note: Forecast datetime will need to be adjusted to make sense in your timezone
+      int forecast_hour                  = hour(x.datetime);
+      int forecast_day_of_week           = weekday(x.datetime);   // the weekday (Sunday is day 1) 
+      int forecast_month                 = month(x.datetime);
+      int forecast_day_of_month          = day(x.datetime);      // the day of month (first day of the month is 1, not 0 :-) )   
+
+      char weekday_str[12] = { '\0' };
+      strcpy(weekday_str, dayStr(forecast_day_of_week));    
+/*
+      char am_or_pm[3];
+      if (isAM(x.datetime) == false)
+      {
+        strcpy(am_or_pm, "PM");
+      }
+      else
+      {
+        strcpy(am_or_pm, "AM");
+      }
+*/      
+      Serial.printf_P("Forecast for %s %d %s @ %d:00: %s and %d degrees.\n", weekday_str, forecast_day_of_month, monthStr(forecast_month), hour(x.datetime), x.summary, x.temp/10);    
+      
+    }    
+
+    delay(10000);
 
 }
